Add SonarCloud workflow and update Java version property
Introduced a new GitHub action workflow to incorporate SonarCloud analysis for every push and pull request made on the 'main' branch. This workflow involves building and analyzing the code on an Ubuntu machine, caching SonarCloud and Maven packages, and getting the Java version from the pom.xml file. Also, updated the pom.xml file to use a 'java.version' property for specifying the Java version instead of directly setting maven compiler source and target. This ensures consistent usage of the Java version across the project.