Add GitHub Packages repository and XPath environment variable
In the build workflow, an environment variable XPATH has been added to make the XML parsing in the jobs less verbose. Meanwhile, in the pom.xml file, a repository and a snapshotRepository were added to configure deployment to GitHub Packages. These changes will streamline builds and make package distribution easier and more efficient.