Add jmp-cloud-service-impl module to Maven POM
Added a new Maven module, 'jmp-cloud-service-impl', to the project. This change was reflected both in the creation of a new pom.xml for the module and adding the module to the modules list in the root pom.xml. This new module is needed to implement cloud-specific services in a separate module to adhere to modularity and clean architecture rules.