Add BankCard, Subscription, User DTOs
Three new data transfer objects (DTOs) are added to keep a clear separation between the database persistence layer and the domain/application layer or service layer. These are 'BankCard', 'Subscription', and 'User'. They contain methods to get properties, equality and hash code methods. This will take data from the database and provide it to the application service layer for processing.

closes #3