Add unit tests, update build dependencies and improve code quality
Introduced unit tests for the classes in the jmp-dto module to ensure they function as expected. Added Spock Framework dependencies to the jmp-cloud-bank-impl and jmp-cloud-service-impl modules to facilitate writing of comprehensive yet maintainable tests, and incorporated the usage of AssertJ and JUnit Jupiter. Modified constructors of the BankCard and Subscription classes to throw NullPointerException if null arguments are provided, upholding robustness principle. Updated the plugins in 'pom.xml' to include Surefire for test execution and reporting. Lastly, incorporated a Quality Gate Status badge in 'README.md' for tracking code quality.