Integrate JaCoCo for improved test coverage
Added support for JaCoCo (Java Code Coverage Library) to the project, which is implemented in the 'pom.xml' file. This plugin will ensure a greater reach of test coverage for our code base, thereby boosting the product's overall quality. To utilize it, specific executions such as 'prepare-agent' and 'report' have been specified, with the report format set to XML.