Replace JUnit tests with Spock, update dependencies
This change replaces the JUnit based unit tests with the Spock testing framework in the Subscription, BankCard and User classes for the jmp-dto project. Spock framework provides more comprehensive and conciser syntax for tests that makes it more maintainable and readable. Additionally, it resolves the dependencies in jmp-cloud-bank-impl's pom.xml by replacing the hardcoded version of jmp-bank-api with the variable reference, which should make the project easier to manage when versions change. Also, edits were made to the execution goals in the root pom.xml to include additional Groovy-specific compile steps, allowing for better integration with the Spock framework.