Updated pom.xml with new plugins and ID changes
In pom.xml, the distributionManagement IDs were changed from 'github' to 'site' and 'repository' respectively for better distinctiveness. Also, the maven-scm-publish-plugin and maven-project-info-reports-plugin were added to provide support for publishing project information and source control management. Lastly, 'HEAD' tag was added to the SCM section and maven-scm-publish-plugin was added to the pluginManagement section for comprehensive SCM configuration.