Refactored packages and updated dependencies
Refactored the packages across the codebase, renaming 'com.epam.engx.jmp' to 'com.epam.jmp'. Removed unnecessary classes and updated dependencies in the pom.xml files with more recent versions. Created new User entity in the jmp-dto module and connected it with the database h2 for testing purposes. This was necessary to make the codebase cleaner, maintain a consistent naming convention and to accommodate new JPA and Spring requirements.