Refactored Controller.java into two separate controllers and added tests
Controller.java was refactored into two separate controllers, BankController.java and ServiceController.java, to improve organization and separation of concerns. The url for creating a card was also adjusted to be more restful by changing "/createCard" to "/card". Additionally, test dependencies were added to pom.xml, and a new test file, BankControllerCreateBankCardTests.java, was added to verify the behavior of createCard operation in BankController. Furthermore, module-info.java was updated to export the controller package.