Add functionality for creating bank cards and subscriptions
This update added functionality to create and manage bank cards, as well as user subscriptions to the service. For the new functionalities, new endpoints were setup with better defined routes following RESTful conventions. Moreover, dependencies required for running tests were added, and tests were written to verify correct behavior of the new functionalities. This is critical for service stability and reliable transactions. Internal implementation was also adjusted accordingly to support the new features.