Refactor isPayableUser method and tests
A new LegalAgePredicate class has been created to replace the MATURITY_AGE constant in the Service class to make the age-checking logic reusable. The ServiceTest class was deleted, a new ServiceAiTest class has been created. This class includes tests for the new predicate. The isPayableUser method in the Service interface has been updated to use the new LegalAgePredicate.